+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add the `refreshCache` param to NFT rarity endpoints to allow users to update stale rarity values.
+
 ### Major Changes
 
 ### Minor Changes