+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add the `refreshCache` parameter to NFT rarity endpoints to allow users to update stale rarity values.
+- Add the `includeCount` parameter to getOwnersForContract.
 
 ### Major Changes