+++ CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## Unreleased
 
-- Add the `refreshCache` parameter to NFT rarity endpoints to allow users to update stale rarity values.
-- Add the `includeCount` parameter to getOwnersForContract.
-
 ### Major Changes
 
 ### Minor Changes
+
+- Add the `refreshCache` parameter to NFT rarity endpoints to allow users to update stale rarity values.
+- Add the `includeCount` parameter to getOwnersForContract.
 
 ## 2.9.0